feat(ITA3): Fixed login implementation using only Jwt tokens
