+++ src/main/java/com/dragonballs/security/WebSecurityConfig.java
@@ -1,6 +1,6 @@
 package com.dragonballs.security;
 
-import com.dragonballs.services.user.JwtAuthorizationTokenFilter;
+import com.dragonballs.services.authentication.JwtAuthorizationTokenFilter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
