+++ src/main/java/com/dragonballs/services/authentication/JwtAuthenticationResponse.java
@@ -1,4 +1,4 @@
-package com.dragonballs.services.user;
+package com.dragonballs.services.authentication;
 
 import java.io.Serializable;
