+++ src/main/java/com/dragonballs/services/authentication/JwtAuthorizationTokenFilter.java
@@ -1,0 +1,53 @@
+package com.dragonballs.services.authentication;
+
+import com.dragonballs.services.authentication.JwtTokenUtil;
+import com.dragonballs.services.user.UserService;
+import io.jsonwebtoken.ExpiredJwtException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Component
+public class JwtAuthorizationTokenFilter extends OncePerRequestFilter {
+
+    @Autowired
+    private UserService userService;
+
+    @Autowired
+    private JwtTokenUtil jwtTokenUtil;
+
+    private final String tokenHeader = "Authorization";
+
+    public JwtAuthorizationTokenFilter() {
+    }
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
+            throws ServletException, IOException {
+
+        final String requestHeader = request.getHeader(this.tokenHeader);
+
+        String username = null;
+        String authToken = null;
+        if (requestHeader != null) {
+            authToken = requestHeader.substring(7);
+            try {
+                username = jwtTokenUtil.getUsernameFromToken(authToken);
+            } catch (IllegalArgumentException e) {
+                logger.error("an error occurred during getting username from token", e);
+            } catch (ExpiredJwtException e) {
+                logger.warn("the token is expired and not valid anymore", e);
+            }
+        } else {
+            logger.warn("couldn't find bearer string, will ignore the header");
+        }
+
+        filterChain.doFilter(request, response);
+    }
+}
