+++ src/main/java/com/dragonballs/services/user/UserService.java
@@ -63,40 +63,4 @@ public class UserService {
     public User getUserByUsername(String username) {
         return userDAO.findByUsername(username);
     }
-
-    /*
-    public User createSessionToken(User user) {
-        User existingUser = userDAO.findByUsername(user.getUsername());
-
-        if (existingUser == null) {
-            throw new UserException("User does not exist");
-        }
-
-        if (!bCryptPasswordEncoder.matches(user.getPasswordHash(), existingUser.getPasswordHash())) {
-            throw new UserException("Password is incorrect");
-        }
-
-        // Is it okay to ignore NoSuchAlgorithmException???
-        try {
-            MessageDigest messageDigest = MessageDigest.getInstance("SHA-256");
-            messageDigest.update(user.getEmail().getBytes());
-            user.setSessionToken(new String(messageDigest.digest()));
-        } catch (NoSuchAlgorithmException ex) {
-            throw new UserException("Something went wrong with authentication");
-        }
-
-        return user;
-    }
-
-    public boolean isUserLoggedIn(String userSessionToken) {
-        User loggedInUser = userDAO.findBySessionToken(userSessionToken);
-
-        if (loggedInUser == null) {
-            return false;
-        }
-
-        return true;
-    }
-    */
-
 }
